rename: CustomButton -> Button 이름 수정
- w3 멘토링 조언 반영
