+++ frontend/src/components/Buttons/index.ts
@@ -1,6 +1,6 @@
-import CustomButton from './CustomButton';
+import Button from './Button';
 import IconButton from './IconButton';
 import IconToggleButton from './IconToggleButton';
 import LogoButton from './LogoButton';
 
-export { CustomButton, IconButton, LogoButton, IconToggleButton };
+export { Button, IconButton, LogoButton, IconToggleButton };
